Move things out of FanoutGraphqlExpressServer and into the beginnings of modules
