+++ src/WebSocketApolloClient.ts
@@ -1,9 +1,8 @@
 import { InMemoryCache } from "apollo-cache-inmemory";
-import { ApolloClient, SubscriptionOptions } from "apollo-client";
+import { ApolloClient } from "apollo-client";
 import { split } from "apollo-link";
 import { createHttpLink } from "apollo-link-http";
 import { WebSocketLink } from "apollo-link-ws";
-import { gql } from "apollo-server";
 import { getMainDefinition } from "apollo-utilities";
 import fetch from "cross-fetch";
 import * as WebSocket from "ws";
