+++ src/examples/apollo-server-express-api.ts
@@ -1,29 +1,58 @@
 /**
  * API from https://www.apollographql.com/docs/apollo-server/features/subscriptions#middleware
  */
-import { ApolloServer, PubSub } from 'apollo-server-express'
-import * as express from "express"
-import * as http from "http"
-import FanoutGraphqlApolloConfig from '../FanoutGraphqlApolloConfig';
-import { MapSimpleTable } from '../SimpleTable';
+import {
+  ApolloServer,
+  buildSchemaFromTypeDefinitions,
+  PubSub,
+} from "apollo-server-express";
+import * as express from "express";
+import * as http from "http";
+import FanoutGraphqlApolloConfig, {
+  FanoutGraphqlTypeDefs,
+} from "../FanoutGraphqlApolloConfig";
+import EpcpPubSubMixin from "../graphql-epcp-pubsub/EpcpPubSubMixin";
+import { MapSimpleTable } from "../SimpleTable";
+import GraphqlWsOverWebSocketOverHttpExpressMiddleware from "../subscriptions-transport-ws-over-http/GraphqlWsOverWebSocketOverHttpExpressMiddleware";
 
 const PORT = process.env.PORT || 4000;
 const app = express();
-const server = new ApolloServer(FanoutGraphqlApolloConfig({
-  pubsub: new PubSub(),
-  subscriptions: true,
-  tables: {
-    notes: MapSimpleTable(),
-  },
-}));
 
-server.applyMiddleware({app})
+// This is what you need to support WebSocket-Over-Http Subscribes
+app.use(GraphqlWsOverWebSocketOverHttpExpressMiddleware());
+
+// This is what you need to support EPCP Publishes (make sure it gets to your resolvers who call pubsub.publish)
+const pubsub = EpcpPubSubMixin({
+  grip: {
+    url: process.env.GRIP_URL || "http://localhost:5561",
+  },
+  // Build a schema from typedefs here but without resolvers (since they will need the resulting pubsub to publish to)
+  schema: buildSchemaFromTypeDefinitions(FanoutGraphqlTypeDefs(true)),
+})(new PubSub());
+
+const apolloServer = new ApolloServer(
+  FanoutGraphqlApolloConfig({
+    pubsub,
+    subscriptions: true,
+    tables: {
+      notes: MapSimpleTable(),
+    },
+  }),
+);
+
+apolloServer.applyMiddleware({ app });
 
 const httpServer = http.createServer(app);
-server.installSubscriptionHandlers(httpServer);
+apolloServer.installSubscriptionHandlers(httpServer);
 
 // ⚠️ Pay attention to the fact that we are calling `listen` on the http server variable, and not on `app`.
 httpServer.listen(PORT, () => {
-  console.log(`🚀 Server ready at http://localhost:${PORT}${server.graphqlPath}`)
-  console.log(`🚀 Subscriptions ready at ws://localhost:${PORT}${server.subscriptionsPath}`)
-})
+  console.log(
+    `🚀 Server ready at http://localhost:${PORT}${apolloServer.graphqlPath}`,
+  );
+  console.log(
+    `🚀 Subscriptions ready at ws://localhost:${PORT}${
+      apolloServer.subscriptionsPath
+    }`,
+  );
+});
