+++ src/index.ts
@@ -1,4 +1,3 @@
-import * as aws from "@pulumi/aws";
 import * as awsx from "@pulumi/awsx";
 import * as pulumi from "@pulumi/pulumi";
 import FanoutGraphqlAwsApp from "./FanoutGraphqlAwsApp";
