--- src/subscriptions-transport-core/utils/empty-iterable.ts
@@ -1,18 +1,0 @@
-import { $$asyncIterator } from "iterall";
-
-export const createEmptyIterable = (): AsyncIterator<any> => {
-  return {
-    next() {
-      return Promise.resolve({ value: undefined, done: true });
-    },
-    return() {
-      return Promise.resolve({ value: undefined, done: true });
-    },
-    throw(e: Error) {
-      return Promise.reject(e);
-    },
-    [$$asyncIterator]() {
-      return this;
-    },
-  };
-};
