--- src/subscriptions-transport-core/utils/is-object.ts
@@ -1,3 +1,0 @@
-export default function isObject(value?: any): boolean {
-  return value !== null && typeof value === "object";
-}
