--- src/subscriptions-transport-core/utils/is-string.ts
@@ -1,3 +1,0 @@
-export default function isString(value?: any): value is string {
-  return typeof value === "string";
-}
