--- src/subscriptions-transport-core/utils/is-subscriptions.ts
@@ -1,10 +1,0 @@
-import { DocumentNode, getOperationAST } from "graphql";
-
-export const isASubscriptionOperation = (
-  document: DocumentNode,
-  operationName?: string,
-): boolean => {
-  const operationAST = getOperationAST(document, operationName);
-
-  return !!operationAST && operationAST.operation === "subscription";
-};
