+++ tslint.json
@@ -23,9 +23,6 @@
     "linterOptions": {
         "exclude": [
             "node_modules/**",
-            "src/subscriptions-transport-ws-over-http/**",
-            "src/subscriptions-transport-core/utils/**",
-            "src/subscriptions-transport-core/legacy/**"
         ]
     }
 }
